impl cvar, move shared gvar code
