+++ read-fonts/src/tables/cvar.rs
@@ -4,12 +4,42 @@
 include!("../../generated/generated_cvar.rs");
 
 use super::variations::{
-    DeltaRunIter, PackedDeltas, PackedPointNumbers, PackedPointNumbersIter, Tuple,
-    TupleVariationCount, TupleVariationHeader, TupleVariationHeaderIter,
+    PackedPointNumbers, TupleVariationCount, TupleVariationData, TupleVariationHeader,
 };
+
+impl<'a> Cvar<'a> {
+    pub fn variation_data(&self, axis_count: u16) -> Result<TupleVariationData<'a>, ReadError> {
+        let count = self.tuple_variation_count();
+        let data = self.data()?;
+        let header_data = self.raw_tuple_header_data();
+        // if there are shared point numbers, get them now
+        let (shared_point_numbers, serialized_data) = if count.shared_point_numbers() {
+            let (packed, data) = PackedPointNumbers::split_off_front(data);
+            (Some(packed), data)
+        } else {
+            (None, data)
+        };
+        Ok(TupleVariationData {
+            is_gvar: false,
+            tuple_count: count,
+            axis_count,
+            shared_tuples: None,
+            shared_point_numbers,
+            header_data,
+            serialized_data,
+        })
+    }
+
+    fn raw_tuple_header_data(&self) -> FontData<'a> {
+        let range = self.shape.tuple_variation_headers_byte_range();
+        self.data.split_off(range.start).unwrap()
+    }
+}
 
 #[cfg(test)]
 mod tests {
+    use font_types::F2Dot14;
+
     use crate::{FontRef, TableProvider};
 
     #[test]
@@ -106,6 +136,19 @@ mod tests {
                 ],
             ),
         ];
-        for (coords, deltas) in cases {}
+        let cvar = font.cvar().unwrap();
+        let axis_count = font.fvar().unwrap().axis_count();
+        let cvar_data = cvar.variation_data(axis_count).unwrap();
+        for (coords, expected_deltas) in cases {
+            let coords = coords.map(F2Dot14::from_f32);
+            let mut deltas = vec![0; expected_deltas.len()];
+            for (tuple, weight) in cvar_data.active_tuples_at(&coords) {
+                for delta in tuple.deltas() {
+                    let scaled_delta = delta.apply_scalar(weight).x;
+                    deltas[delta.position as usize] += scaled_delta.to_bits();
+                }
+            }
+            assert_eq!(&deltas, expected_deltas);
+        }
     }
 }
